Add Algorithm model and refactor base HMAC logic
