+++ Sources/Uno/CounterBasedGenerator.swift
@@ -25,12 +25,15 @@ public struct CounterBasedGenerator: AuthenticationCodeGenerator {
   public let secret: Secret
 
   public let codeLength: Int
+  
+  public let algorithm: Algorithm
 
   // MARK: - Init
   
-  public init(secret: Secret, codeLength: Int = 6) {
+  public init(secret: Secret, codeLength: Int = 6, algorithm: Algorithm = .sha1) {
     self.secret = secret
     self.codeLength = codeLength
+    self.algorithm = algorithm
   }
 }
 
@@ -42,16 +45,16 @@ extension CounterBasedGenerator {
   ///   - counter: A variable number that acts as a seed for the generator.
   /// - Returns: A `String` representing the generated one-time password.
   public func generate(from counter: UInt64) throws -> String {
-    let hash = try generateHMACHash(from: counter)
-    let code = hash.dynamicallyTrimmed(numberOfDigits: codeLength)
+    let hmac = try generateHMAC(from: counter)
+    let code = hmac.dynamicallyTrimmed(numberOfDigits: codeLength)
     return code
   }
   
   /// Generates the HMAC hash from a secret payload and a moving factor.
   /// - Parameters:
   ///   - counter: A variable number that acts as a seed for the generator.
-  /// - Returns: A `MessageAuthenticationCode` representing the generated hash.
-  func generateHMACHash(from counter: UInt64) throws -> some MessageAuthenticationCode {
+  /// - Returns: A `Data` payload representing a HMAC.
+  func generateHMAC(from counter: UInt64) throws -> Data {
     guard isCodeLengthValid else {
       throw Error.codeLengthNotSupported
     }
@@ -59,6 +62,27 @@ extension CounterBasedGenerator {
     var counter = counter.bigEndian
     let counterData = Data(bytes: &counter, count: MemoryLayout<UInt64>.size)
     
-    return HMAC<Insecure.SHA1>.authenticationCode(for: counterData, using: secret.symmetricKey)
+    return generateHMAC(from: secret.symmetricKey, data: counterData)
+  }
+  
+  /// Generates the HMAC from a secret key and data payload.
+  /// - Note: Implementation works around a limitation for protocols with associated types
+  /// by returning a `Data` instead of a more specific `MessageAuthenticationCode`.
+  /// This issue might be fixed in future implementations.
+  /// - Parameters:
+  ///   - key: The symmetric cryptographic key to use as seed.
+  ///   - data: The data payload to use as seed.
+  /// - Returns: A `Data` payload representing a HMAC.
+  func generateHMAC(from key: SymmetricKey, data: Data) -> Data {
+    switch algorithm {
+      case .sha1:
+        return Data(HMAC<Insecure.SHA1>.authenticationCode(for: data, using: key))
+        
+      case .sha256:
+        return Data(HMAC<SHA256>.authenticationCode(for: data, using: key))
+
+      case .sha512:
+        return Data(HMAC<SHA512>.authenticationCode(for: data, using: key))
+    }
   }
 }
